Undo renames to fix diff
